consolidate tests into stream_iter.rs
